fake daemonize() on windows for now
Not sure how to actually implement it, maybe we have to look into
Windows Services.
