+++ subcommands/agent/daemonize.go
@@ -1,0 +1,34 @@
+//go:build !windows
+
+package agent
+
+import (
+	"fmt"
+	"os"
+	"syscall"
+)
+
+func daemonize(argv []string) error {
+	binary, err := os.Executable()
+	if err != nil {
+		return err
+	}
+
+	procAttr := syscall.ProcAttr{}
+	procAttr.Files = []uintptr{
+		uintptr(syscall.Stdin),
+		uintptr(syscall.Stdout),
+		uintptr(syscall.Stderr),
+	}
+	procAttr.Env = append(os.Environ(),
+		"REEXEC=1",
+	)
+
+	pid, err := syscall.ForkExec(binary, argv, &procAttr)
+	if err != nil {
+		return err
+	}
+	fmt.Printf("agent started with pid=%d\n", pid)
+	os.Exit(0)
+	return nil
+}
