+++ subcommands/agent/daemonize_windows.go
@@ -1,0 +1,9 @@
+//go:build windows
+
+package agent
+
+import "errors"
+
+func daemonize(argv []string) error {
+	return errors.ErrUnsupported
+}
